+++ examples/automator/swatcher_quick_action.py
@@ -1,10 +1,11 @@
 import sys
 from swatcher import Swatcher
 
-files = sys.argv[1:]
 
-for file in files:
-    s = Swatcher(file)
-    s.sample(max_colors=10)
-    s.export_palette_image()
-    s.export_ase_file()
+if __name__ == "__main__":
+    files = sys.argv[1:]
+
+    for file in files:
+        s = Swatcher(file)
+        s.export_ase_file()
+        s.export_palette_image()