+++ swatcher/image.py
@@ -22,30 +22,23 @@ def trim_excess(image: object) -> object:
     # count how many times each value is present
     color_count = Counter([pixel for pixel in corners]).most_common()
 
-    # use the top left pixel as a base case
-    bg_pixel = image.getpixel((0, 0))
-
-    # if four corners don't match
-    if len(color_count) > 1:
-        # and multiple corners don't have the same count
-        if color_count[0][1] != color_count[1][1]:
-            # set the comparison pixel to the most common value
-            bg_pixel = color_count[0][0]
-        else:
-            # if two corners have the same count don't trim
-            return image
+    # if multiple corners have the same pixel count don't trim
+    if len(color_count) > 1 and color_count[0][1] == color_count[1][1]:
+        return image
+    else:  # set the comparison pixel to the most common value
+        bg_pixel = color_count[0][0]
 
     # compare the original image to the excess pixels
-    comparison = Image.new("RGB", image.size, bg_pixel)
-    diff = ImageChops.difference(image, comparison)
+    comp = Image.new("RGB", image.size, bg_pixel)
+    diff = ImageChops.difference(image, comp)
     bbox = diff.getbbox()
-    # crop the image to the difference
+    # crop the difference
     return image.crop(bbox)
 
 
 def process_image(image: object, max_size: int = 500) -> object:
     """
-    Flatten, Crop, and Scale the provided image.
+    Process the image for best color sampling results.
 
     :param image: PIL Image object
     :param max_size: maximum size of the image for color sampling
@@ -58,15 +51,14 @@ def process_image(image: object, max_size: int = 500) -> object:
         raise ValueError("The provided image has no pixels.")
 
     # composite the image on a white background just in case it has transparency
-    background = Image.new("RGBA", image.size, (255, 255, 255))
-    composite = Image.alpha_composite(background, image)
+    bg = Image.new("RGBA", image.size, (255, 255, 255))
+    comp = Image.alpha_composite(bg, image)
     # convert composite image to RGB since we only need the RGB color values
-    composite = composite.convert("RGB")
+    comp = comp.convert("RGB")
     # crop the image if extra surrounding background pixels are found
-    composite = trim_excess(composite)
+    comp = trim_excess(comp)
     # reduce the image down to `max_size` to speed up processing
-    if composite.width > max_size or composite.height > max_size:
-        new_size = (max_size, max_size)
-        composite.thumbnail(new_size, resample=0)
+    if comp.width > max_size or comp.height > max_size:
+        comp.thumbnail((max_size, max_size), resample=0)
 
-    return composite
+    return comp